+++ CHANGELOG.md
@@ -42,6 +42,10 @@ All notable changes to the LoLLMs Platform will be documented in this file.
 
 - feat(notebook): add tab support and extend creation schema
 
+## [2026-09-06 21:41]
+
+- # Commit message for frontend asset deletions and backend minor tweaks
+
 ## [2026-09-04 12:26]
 
 - build(deps): bump lollms-client from 1.18.2 to 1.18.4